cams

+++ app/models/security_camera.rb
@@ -51,26 +51,33 @@ class SecurityCamera < ApplicationRecord
         puts tmp_path
         begin
             puts "[open] #{self.image_url}"
-            open(self.image_url) {|f|
-            File.open(tmp_path,"wb") do |file|
-                file.puts f.read
-            end
-            }
+            # open(self.image_url) {|f|
+            # File.open(tmp_path,"wb") do |file|
+            #     file.puts f.read
+            # end
+            # }
+            wget_command="wget #{self.image_url} -O #{tmp_path}"
+            puts wget_command
+            if system(wget_command)&&File.file?(tmp_path)
 
-            final_path="#{self.camera_image_dir}/#{DateTime.now.to_i.to_s}.jpg"
-            convert_command="convert #{tmp_path} -resize 1920x1080^ -gravity center -quality 75 #{final_path}"
-            puts "[convert] #{convert_command}"
-            if system(convert_command)
-                return false 
-                self.update_attribute(:last_seen,DateTime.now)
-                FileUtils.rm(tmp_path)
-                REDIS.del("screenshot:timer")
-                self.update_attribute(:error_count,0)  
-    
-                return true
+                final_path="#{self.camera_image_dir}/#{DateTime.now.to_i.to_s}.jpg"
+                convert_command="convert #{tmp_path} -resize 1920x1080^ -gravity center -quality 75 #{final_path}"
+                puts "[convert] #{convert_command}"
+                if system(convert_command)
+                    self.update_attribute(:last_seen,DateTime.now)
+                    FileUtils.rm(tmp_path)
+                    REDIS.del("screenshot:timer")
+                    self.update_attribute(:error_count,0)  
+        
+                    return true
+                else
+                    self.update_attribute(:error_count,self.error_count+1)  
+                    puts "[convert] command failed"
+                    return false 
+
+                end
             else
-                self.update_attribute(:error_count,self.error_count+1)  
-                puts "[convert] command failed"
+                puts "[wget] fails"
             end
         rescue => e          
             self.update_attribute(:error_count,self.error_count+1)
